integrate value converter and observer dump

+++ bark_ml/evaluators/base_evaluator.hpp
@@ -48,11 +48,11 @@ using bark::models::dynamic::StateDefinition::X_POSITION;
 using bark::models::dynamic::StateDefinition::Y_POSITION;
 using bark::models::dynamic::StateDefinition::THETA_POSITION;
 using bark::models::dynamic::StateDefinition::VEL_POSITION;
-using ObservedState = Eigen::Matrix<float, 1, Eigen::Dynamic>;
+using ObservedState = Eigen::Matrix<double, 1, Eigen::Dynamic>;
 using bark::commons::transformation::FrenetPosition;
-using State = Eigen::Matrix<float, Eigen::Dynamic, 1>;
+using State = Eigen::Matrix<double, Eigen::Dynamic, 1>;
 
-using Reward = float;
+using Reward = double;
 using Done = bool;
 using EvalResults = std::map<std::string, EvaluationReturn>;
